+++ twitter/friends.go
@@ -57,7 +57,7 @@ func (s *FriendService) IDs(params *FriendIDParams) (*FriendIDs, *http.Response,
 type FriendListParams struct {
 	UserID              int64  `url:"user_id,omitempty"`
 	ScreenName          string `url:"screen_name,omitempty"`
-	Cursor              int    `url:"cursor,omitempty"`
+	Cursor              int64  `url:"cursor,omitempty"`
 	Count               int    `url:"count,omitempty"`
 	SkipStatus          *bool  `url:"skip_status,omitempty"`
 	IncludeUserEntities *bool  `url:"include_user_entities,omitempty"`
